refactor: logical directory restructure
