+++ common/src/lib.rs
@@ -2,7 +2,5 @@ extern crate lazy_static;
 
 pub mod constants;
 pub mod ether;
-pub mod io;
 pub mod resources;
-pub mod testing;
 pub mod utils;
