+++ common/src/resources/transpose.rs
@@ -3,7 +3,7 @@ use reqwest::header::HeaderMap;
 use serde_json::Value;
 use std::time::{Duration, Instant};
 
-use crate::io::logging::Logger;
+use crate::utils::io::logging::Logger;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
