+++ common/src/utils/mod.rs
@@ -1,7 +1,9 @@
 pub mod http;
 pub mod integers;
+pub mod io;
 pub mod strings;
 pub mod sync;
+pub mod testing;
 mod tests;
 pub mod threading;
 pub mod time;
