+++ core/src/cfg/mod.rs
@@ -11,7 +11,7 @@ use clap::{AppSettings, Parser};
 use heimdall_common::{
     constants::{ADDRESS_REGEX, BYTECODE_REGEX},
     ether::evm::core::vm::VM,
-    io::logging::*,
+    utils::io::logging::*,
 };
 use petgraph::Graph;
