+++ core/src/decompile/resolve.rs
@@ -1,5 +1,5 @@
 use super::util::Function;
-use heimdall_common::{ether::signatures::ResolvedFunction, io::logging::Logger};
+use heimdall_common::{ether::signatures::ResolvedFunction, utils::io::logging::Logger};
 
 // match the ResolvedFunction to a list of Function parameters
 pub fn match_parameters(
