+++ core/src/dump/util/csv.rs
@@ -2,9 +2,9 @@ use ethers::{
     abi::{decode, ParamType},
     types::U256,
 };
-use heimdall_common::{
+use heimdall_common::utils::{
     io::file::write_lines_to_file,
-    utils::strings::{encode_hex, hex_to_ascii},
+    strings::{encode_hex, hex_to_ascii},
 };
 
 use crate::dump::{constants::DECODE_AS_TYPES, structures::dump_state::DumpState};
