+++ core/tests/test_cfg.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod benchmark {
     use clap_verbosity_flag::Verbosity;
-    use heimdall_common::testing::benchmarks::async_bench;
+    use heimdall_common::utils::testing::benchmarks::async_bench;
 
     use heimdall_core::cfg::CFGArgs;
